Updated tuf_interface.c  Now compiles and runs
I changed the PyObject_CallObject to a PyObject_CallFunction, the
code now compiles, runs, and *doesn't* seg fault! /cheer.  This is a
step forward at least... We still need to figure out how to play with
the return values, but we can now at the very least cleanly call python
functions inside of C.

NOTE:  Make sure you to use Py_XDECREF() as opposed to Py_DECREF.  The X
makes use of a different Macro, and if the point is NULL, the X Macro
will do nothing, where os the DECREF macro will try to remove a NULL
reference, thus inducing a seg fault :P