Created a working repository for our tests
Created a functional repository that contains sample files that can
be used for 'gem install' and 'gem update'. When installing or
updating a gem, the following files are fetched from the repo:
   /specs.4.8.gz
   /quick/Marshal.4.8/gemname-*.*.*.gemspec.rz
   /gems/gemname-*.*.*.gem

Such files are now included in the repo along with proper TUF
metadata. Thus, TUF integration into gems can be tested.