Updated the urllib functions
Tony figured out that the issues regarding the URLLIB functions had to
do with our memory reference counters.  Super made props to him.  The
updated interface now allows for us to call urllib_tuf.urlopen,
urllib2_tuf.urlopen, and urllib_tuf.urlretrieve, but we haven't been
able to get all functions to be called simultaneously.

I don't believe this is a very big issue however.  When we're using this
interface from within ruby, these functions will be called on their own.
Our next big challenge is to ensure that the return values for each
function are what we need in Ruby to finish bridging the gap!