Added Makefile to install gem that uses TUF proxy
Using the given Makefile and the files inside proxy/ directory
we can install a version of rubygems that uses TUF through a proxy.

In order to test, first enable python virtualenv and rvm:
	source virtualpython/bin/activate
	source ~/.rvm/scripts/rvm

Make sure you are using Ruby 2.0.0 in rvm:
	rvm use 2.0.0

Then use the given Makefile to install web.py, TUF and Rubygems:
	make

The Makefile uses pip (inside python virtualenv) to install web.py
and TUF, and the uses git clone to download the latest Rubygems.
In addition, the Makefile copies necessary files to the bin/ directory
of gem inside RVM (i.e. the current gem that we are using) as
well as to the home folder of the current user (/home/username)

Finally, to test proper operation, first start our local TUF repo:
	./myserver.rb

and then try to use gem as usual:
	gem search
	gem install nekgem -v "0.0.3"
	gem update nekgem
	gem uninstall -a nekgem