Moved scratch work to ./dev, and rm'd rvm.sh
I did some general cleanup of the repo.  From now on let's do all of our
testing work under the ./dev folder.  Once we can verify with certainty
that the changes we make are stable we can push them to our 'production'
folder which is the 'bridge' folder in the root.